requirements.txt
pyyaml
markdown
jinja2
Pillow